+++ src/npc/npc.py
